inserting information clues in links.txt
These information lines would help at separating the information in the unified json file according to their originators. I propose the actual format but it can be really anything that wouln't trigger citron to look for information. Keeping the same separation clue - as in "www" from the links- would help towards automation of the information processing.